Fix programmatically initial text not formatted
Now when we start watching to sync the number with the flag, we first check if there’s already a valid number in the input. If there is, we ignore the initial country code and go with whatever is there — this case will happen when a differnet input set programmatically (e.g. databinding) differs from the initial input.